Move Router to router to follow lowercase mixin convention. (#18)
