+++ package.json
@@ -24,9 +24,9 @@
     "router"
   ],
   "scripts": {
-    "test": "jest --coverage --no-cache --testURL https://hyperapp.js.org",
+    "test": "jest --coverage --no-cache",
     "build": "npm run bundle && npm run minify",
-    "bundle": "rollup -i src/index.js -o dist/router.js -f umd -mn Router",
+    "bundle": "rollup -i src/index.js -o dist/router.js -f umd -mn router",
     "minify": "uglifyjs dist/router.js -o dist/router.js --mangle --compress warnings=false --pure-funcs=Object.defineProperty -p relative --source-map dist/router.js.map",
     "prepublish": "npm run build",
     "format": "prettier --semi false --write 'src/**/*.js'",
