+++ src/index.js
@@ -1,2 +1,2 @@
-export { Router } from "./Router"
+export { router } from "./router"
 export { Link } from "./Link"
